refactor(ios): decompose IosGattServer 684->390 loc via extension functions
Split IosGattServer.kt into three focused files following the
Class Decomposition Pattern, Approach B (Extension Functions):

- IosGattServer.kt (390 loc): facade with GattServer interface impl,
  state fields (internal), open/notify/indicate/close API,
  setDelegateCallbacks, trackCentral, runSweepLoop, sendUpdate

- IosGattServerExtensions.kt (113 loc): file-level constants
  (POWER_ON_TIMEOUT_MS, MAX_NOTIFY_RETRIES, etc.), CBATTRequest/
  CBCentral extension properties, buildCBProperties/buildCBPermissions,
  buildNativeService (stand-alone), GattStatus.toCBATTError()

- IosGattServerHandlers.kt (215 loc): delegate callback handlers
  as extension functions (handleServiceAdded, handleReadRequest,
  handleWriteRequests, dispatchWrites, assembleFragmentedWrites,
  handleSubscribe, handleReadyToUpdate)
